Adding boneMatrices to MeshUnmanaged to fix double free
I noticed a bug when trying out the models_first_person_maze
example, that sometimes a double free occurs on UnloadMesh
when destructor of MeshUnmanaged is called. The offending line
is Line 1935 from rmodels.c from raylib, where
RL_FREE(mesh.boneMatrices) is called. To remedy this, I added
a default assignment of nullptr to boneMatrices and explicit
assignment throught the set method.